# PERF: speed up PeriodArray creation by exposing dayfirst/yearfirst params