@pep8speaks 2018-12-05
Hello @qwhelan! Thanks for submitting the PR.

 - In the file [`asv_bench/benchmarks/period.py`](https://github.com/qwhelan/pandas/blob/3228823077de49df08627fd029e0d61531297b6d/asv_bench/benchmarks/period.py), following                are the PEP8 issues :
> [Line 38:1](https://github.com/qwhelan/pandas/blob/3228823077de49df08627fd029e0d61531297b6d/asv_bench/benchmarks/period.py#L38): [E302](https://duckduckgo.com/?q=pep8%20E302) expected 2 blank lines, found 1