@jorisvandenbossche 2018-12-06
> Fine by me. Elsewhere we're trying to move towards fewer arguments in the constructors. Should we consider a dedicated constructor for this? (not necessarily for this PR)

I am personally -1 on adding this to the Period constructor. 
(They are also not present on the Timestamp constructor. They are on DatetimeIndex, but I would rather see them deprecated there, in favor of `to_datetime` for all non-default string parsing).

I also find it strange that a `display` option actually determines how an input is parsed. I would rather deprecate that ability (as I understand this is what makes it slower?).