@qwhelan 2018-12-07
> I also find it strange that a display option actually determines how an input is parsed. I would rather deprecate that ability (as I understand this is what makes it slower?).

Yeah, we're spending 90%+ of runtime importing/querying these display variables. Unfortunately, this same code path is used for handling `DatetimeIndex` slicing, so deprecation is likely to be tricky.

The caveat on the above is that this speedup is only when creating a `PeriodArray` directly from integers. This was not covered by our `asv` suite until the past week, but is leveraged in places like `get_indexer_non_unique()`.

Regarding the name of the option, it seems the intent was for it to cover display + parsing, but only parsing was ever implemented (per https://github.com/pandas-dev/pandas/issues/11501#issuecomment-152848626 )