@jreback 2018-12-27
@qwhelan look at how the import is deferred in ``pandas.io.pytables``, e.g. ``_tables()``. not super petty but should work here. basically

```
_get_option = None



def get_option():
     global _get_option
     if _get_option is None:
          from pandas.core.config_init import get_option
          _get_option = get_option
     return _get_option
```